SectionGroup appointments should be floats

+++ src/main/resources/db/migration/V176__Add_reader_and_ta_planned_amounts_to_SectionGroups.sql
@@ -1,2 +1,2 @@
-ALTER TABLE `SectionGroups` ADD COLUMN `TeachingAssistantAppointments` int(11) NULL;
-ALTER TABLE `SectionGroups` ADD COLUMN `ReaderAppointments` int(11) NULL;
+ALTER TABLE `SectionGroups` ADD COLUMN `TeachingAssistantAppointments` FLOAT NULL;
+ALTER TABLE `SectionGroups` ADD COLUMN `ReaderAppointments` FLOAT NULL;
